Implement server-side validation for edit and delete markers. Create public markers query to pull all public markers for main page.

+++ server/routes/markersRoutes.js
@@ -13,8 +13,18 @@ router.get('/category', (req, res) => {
     .catch(err => res.send(err));
 });
 
+router.get('/public/category', (req, res) => {
+  const category = req.query.category;
+  markersQueries.getPublicMarkersByCategory(category)
+    .then(markers => {
+      res.send(markers);
+    })
+    .catch(err => res.send(err));
+});
+
 router.get('/user', (req, res) => {
   const user_id = req.session.userId;
+  console.log(user_id);
   markersQueries.getUserMarkers(user_id)
     .then(markers => {
       res.send(markers);
@@ -50,27 +60,44 @@ router.post('/', (req, res) => {
 
 router.post('/delete/:id', (req, res) => {
   const marker_id = req.params.id;
-  markersQueries.deleteMarker(marker_id)
-    .then(list => {
-      if (list.length < 1) {
-        res.send({error: "error, marker id does not exist"});
-        return;
+  markersQueries.getMarkerById(marker_id)
+    .then((marker) => {
+      if (marker.user_id !== req.session.userId) {
+        console.log("Not the correct user.");
+        res.redirect('../../');
+      } else {
+        markersQueries.deleteMarker(marker_id)
+        .then(list => {
+          if (list.length < 1) {
+            res.send({error: "error, marker id does not exist"});
+            return;
+          }
+          console.log("deleted");
+          res.redirect('../../');
+        })
+        .catch(err => res.send(err));
       }
-      console.log("deleted");
-      res.redirect('../../');
     })
-    .catch(err => res.send(err));
+    .catch(err => {console.log(err)});
 });
 
 router.post('/:id', (req, res) => {
   const marker = req.body;
   marker.id = req.params.id;
-  markersQueries.editMarker(marker, req.session.userId)
-    .then(marker => {
-      console.log("edited");
-      res.redirect('/');
+  markersQueries.getMarkerById(marker.id)
+    .then(mk => {
+      if (mk.user_id !== req.session.userId) {
+        res.redirect('/');
+      } else {
+        markersQueries.editMarker(marker, req.session.userId)
+        .then(marker => {
+          console.log("edited");
+          res.redirect('/');
+        })
+        .catch(err => res.send(err));
+      }
     })
-    .catch(err => res.send(err));
+    .catch(err => {console.log(err)});
 });
 
 module.exports = router;
